Refactor recipe resolution to not just be a Strategy
This allows recipe loading and instantiating, which requires resolution, to not
require Strategies or the Strategizer.
